+++ routes/posts.js
@@ -1,47 +1,40 @@
 const express = require('express');
 const router = express.Router()
 const Posts = require('../models/posts');
-const Comments = require('../models/comments');
+const { protect } = require('../middleware/users');
+
 router.get('/', async (req, res) => {
-  try{
-    const posts = await Posts.find()
+  try {
+    const posts = await Posts.find().populate('userId', 'username').sort({ createdAt: -1 })
+
     res.json(posts)
   }
-  catch (err){
-    res.status(500).json({message: err.message})
+  catch (err) {
+    res.status(500).json({ message: err.message })
   }
 })
-router.get('/comments/:id', async (req, res) => {
-  try{
-    const comments = await Comments.find({postId: req.params.id})
-    res.json(comments)
-  }
-  catch(error){
-    res.status(500).json({message: error.message})
-    
-  }
-})
+
 router.get('/:id', getPost, (req, res) => {
   res.json(res.post)
 })
 
-router.post('/', async (req, res) => {
+router.post('/',protect, async (req, res) => {
   const post = new Posts({
-    userId: req.body.userId,
+    userId: req.user._id,
     postContent: req.body.postContent,
-    
+
   })
-  
-  try{
+
+  try {
     const newPost = await post.save()
     res.status(201).json(newPost)
   }
-  catch(err){
-    res.status(400).json({message: err.message})
+  catch (err) {
+    res.status(400).json({ message: err.message })
   }
 })
 
-router.patch('/:id', getPost, async (req, res) => {
+router.patch('/:id', protect, getPost, async (req, res) => {
   console.log('BODY IS:', req.body)
   if (req.body.postContent == null) {
     return res.status(400).json({
@@ -52,11 +45,16 @@ router.patch('/:id', getPost, async (req, res) => {
   res.post.postContent = req.body.postContent
 
   try {
-    const updatedPost = await res.post.save()
-    res.status(200).json({
-      message: "Post updated.",
-      post: updatedPost
-    })
+    if (req.user._id.equals(res.post.userId._id)) {
+      const updatedPost = await res.post.save()
+      res.status(200).json({
+        message: "Post updated.",
+        post: updatedPost
+      })
+    }
+    else{
+      return res.status(401).json({ "message": `User not authorized.` })
+    }
   } catch (err) {
     res.status(400).json({ message: err.message })
   }
@@ -64,26 +62,31 @@ router.patch('/:id', getPost, async (req, res) => {
 
 
 
-router.delete('/:id', getPost, async (req, res) => {
+router.delete('/:id', protect, getPost, async (req, res) => {
   try {
-    await res.post.deleteOne()
-    res.status(200).json({"message": "Deleted post."})
+    if (req.user._id.equals(res.post.userId._id)) {
+      await res.post.deleteOne()
+      return res.status(200).json({ "message": "Deleted post." })
+    }
+    else {
+      return res.status(401).json({ "message": `User not authorized.` })
+    }
   } catch (err) {
-    res.status(500).json({"message": err.message})
-    
+    res.status(500).json({ "message": err.message })
+
   }
 })
 
-async function getPost(req, res, next){
+async function getPost(req, res, next) {
   let post;
-  try{
-    post = await Posts.findById(req.params.id)
-    if(post === null){
-      return res.status(404).json({"message": "Post not found."})
+  try {
+    post = await Posts.findById(req.params.id).populate('userId', 'username')
+    if (post === null) {
+      return res.status(404).json({ "message": "Post not found." })
     }
   }
-  catch(error){
-    return res.status(500).json({"message": "Server error."});
+  catch (error) {
+    return res.status(500).json({ "message": "Server error." });
   }
   res.post = post
   next()
